+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### next
 
+  * tpm2_flushcontext:
+      - Added option **--cphash** to output the cpHash for the command
+        TPM2_CC_FlushContext.
+
   * tpm2_load:
       - Added capability to load pem files in TSS2-Private-Key format for
         interoperability with tpm2-tss-engine, tpm2-openssl provider,